server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-openai==0.3.13
langchain-text-splitters==0.3.8
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.75.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
propcache==0.3.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
google-generativeai==0.5.4
langchain-google-genai==1.0.5